Renamed and changed behavior of the 'importprivkey' functional test to 'restorewallet'.
